added combat_player node
added a combat_player node that keeps track of the player's shield and counter. added 2 signals to combat manager, one that adds shield, one that adds counter. both are sent out within the process player attack function. the player node sees this and keeps track accordingly. the player node also has a max counter variable that it checks whenever the player's counter increases, and if it reaches it, the player sends out a signal saying that it's counter is ready. (it also resets teh counter to 0, but not sure whether it should do that or allow multiple counters to be readied at once)